[refactor] AddGroup 이름 변경
- AddGroupXXX의 패키지와 클래스명을 SaveGroupXXX로 변경함.
