+++ mobile/app/src/main/java/com/lacuc/pets/ui/ManageModule.kt
@@ -5,9 +5,9 @@ import com.lacuc.pets.ui.manage.animal.choose.ChooseAnimalFragment
 import com.lacuc.pets.ui.manage.animal.detail.AnimalDetailFragment
 import com.lacuc.pets.ui.manage.animal.detail.medical.AddMedicalFragment
 import com.lacuc.pets.ui.manage.animal.detail.memo.AddMemoFragment
-import com.lacuc.pets.ui.manage.group.add.AddGroupFragment
 import com.lacuc.pets.ui.manage.group.choose.ChooseGroupFragment
 import com.lacuc.pets.ui.manage.group.info.UserProfileFragment
+import com.lacuc.pets.ui.manage.group.save.SaveGroupFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -17,7 +17,7 @@ abstract class ManageModule {
     abstract fun getChooseGroupFragment(): ChooseGroupFragment
 
     @ContributesAndroidInjector
-    abstract fun getAddGroupFragment(): AddGroupFragment
+    abstract fun getSaveGroupFragment(): SaveGroupFragment
 
     @ContributesAndroidInjector
     abstract fun getChooseAnimalFragment(): ChooseAnimalFragment
