+++ mobile/app/src/main/res/layout/fragment_save_group.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.manage.group.add.AddGroupFragment">
+    tools:context=".ui.manage.group.save.SaveGroupFragment">
 
     <data>
 
@@ -10,7 +10,7 @@
 
         <variable
             name="vm"
-            type="com.lacuc.pets.ui.manage.group.add.AddGroupViewModel" />
+            type="com.lacuc.pets.ui.manage.group.save.SaveGroupViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
